refactor and fix dependencies
