+++ tests/test_s3.py
@@ -1,53 +1,23 @@
 # -*- coding:utf-8 -*-
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-import io
+import pytest
 
-from treehugger.s3 import _split_s3_filename, include_remote_yaml_data_or_die, read_file_from_s3
-
-
-def test_include_remote_yaml_data_or_die(s3_stub):
-
-    s3_stub.add_response(
-        'get_object',
-        expected_params={
-            'Bucket': 'bucket',
-            'Key': 'filename.yml',
-            'VersionId': '1',
-        },
-        service_response={
-            'Body': io.BytesIO(b'X: y\nZ: zzz'),
-        }
-    )
-    input_yaml_data = {
-        'A': 'b',
-        'include': 's3://bucket/filename.yml?versionId=1'
-    }
-    assert include_remote_yaml_data_or_die(input_yaml_data) == {
-        'A': 'b',
-        'X': 'y',
-        'Z': 'zzz',
-    }
+from treehugger.s3 import split_s3_url
 
 
-def test_split_s3_filename():
-    test_s3_filename = 's3://bucket/key?versionId=7'
-    assert _split_s3_filename(test_s3_filename) == ('bucket', 'key', '7')
+def test_split_s3_url():
+    test_s3_url = 's3://bucket/key?versionId=7'
+    assert split_s3_url(test_s3_url) == ('bucket', 'key', '7')
 
 
-def test_read_file_from_s3(s3_stub):
+def test_no_version_split_s3_url():
+    test_s3_url = 's3://bucket/key'
+    with pytest.raises(SystemExit):
+        split_s3_url(test_s3_url)
 
-    s3_stub.add_response(
-        'get_object',
-        expected_params={
-            'Bucket': 'bucket',
-            'Key': 'filename.yml',
-            'VersionId': '1',
-        },
-        service_response={
-            'Body': io.BytesIO(b'TREEHUGGER_APP: Test\nTREEHUGGER_STAGE: prod\n'),
-        }
-    )
 
-    filename = 's3://bucket/filename.yml?versionId=1'
-    assert read_file_from_s3(filename) == b'TREEHUGGER_APP: Test\nTREEHUGGER_STAGE: prod\n'
+def test_incorrect_scheme_split_s3_url():
+    test_s3_url = 'https://bucket/key?versionId=7'
+    with pytest.raises(SystemExit):
+        split_s3_url(test_s3_url)
