+++ treehugger/yaml.py
@@ -7,7 +7,7 @@ import six
 import yaml
 
 from .messaging import die
-from .s3 import read_file_from_s3
+from .s3 import fetch_s3_content
 
 
 def safe_load(fp_or_text):
@@ -17,7 +17,7 @@ def safe_load(fp_or_text):
 
 def load_file_or_die(filename):
     if filename.startswith('s3://'):
-        return safe_load(read_file_from_s3(filename))
+        return safe_load(fetch_s3_content(filename))
     try:
         with open(filename, 'r') as fp:
             return safe_load(fp)
@@ -56,3 +56,18 @@ def all_strs_text(obj):
         return {six.text_type(k): all_strs_text(v) for k, v in six.iteritems(obj)}
     else:
         return obj
+
+
+def include_remote_yaml_data_or_die(data):
+    """
+    Check for "include" key with URL(s), if found fetch & include the yaml
+    """
+    include_key = 'include'
+    if include_key in data:
+        url = data.pop(include_key)
+        if not isinstance(url, six.string_types):
+            die("'include' value should be a string")
+        yaml_str = fetch_s3_content(url)
+        yaml_data = yaml.safe_load(yaml_str)
+        data.update(yaml_data)
+    return data
